Update arguments for `threshold_summary`, rbuildignore, and cran-comments
This update adds the argument `label_y_axis` which allows to user to
specify whether or not this label should have a y-axis label.

It is helpful to exclude the y-axis label when the parameter being
plotted is not a standard SWMP variable.

Updated rbuildignore to include LICENSE.txt

Updated cran-comments.md to reflect the upgrade to R 3.5.2